move TestGatherLayer_BenchmarkModule to its own file
