+++ tests/utils/test_dist__gather__losses.py
@@ -61,7 +61,6 @@ def close_torch_distributed() -> Generator[None, None, None]:
     torch.distributed.destroy_process_group()
 
 
-@pytest.mark.skip(reason="This test is flaky and needs to be fixed.")
 class TestGatherLayer_Losses:
     """
     Tests that the gather layer works as expected.
@@ -70,7 +69,7 @@ class TestGatherLayer_Losses:
     times, running a proper test is difficult. The approach used here:
 
     1. This test was run once with n_devices=1 and gather=False. The resulting
-    parameters after 10 epochs are saved as expected_params__10_epochs__no_gather.
+    parameters after 10 epochs are hardcoced as expected_params__10_epochs__no_gather.
 
     2. This test is now run with n_devices=2 and gather=True. The resulting
     parameters are asserted to be the same ad with n_devices=1 and gather=False.
@@ -172,63 +171,3 @@ class TestGatherLayer_Losses:
 
         params = list(model.parameters())[0]
         assert torch.allclose(params, expected_params__10_epochs__no_gather, rtol=1e-3)
-
-
-class TestGatherLayer_BenchmarkModule:
-    def test__benchmark_module(self, close_torch_distributed: None) -> None:
-        n_devices = 2
-        n_samples = 128
-        num_classes = 3
-        batch_size = int(n_samples / n_devices)
-
-        pl.seed_everything(0, workers=True)
-
-        dataset_train = LightlyDataset.from_torch_dataset(
-            FakeData(
-                size=n_samples,
-                image_size=(3, 32, 32),
-                num_classes=num_classes,
-                transform=ToTensor(),
-            )
-        )
-        dataloader_train = DataLoader(
-            dataset_train,
-            batch_size=batch_size,
-            num_workers=0,
-            shuffle=False,
-            drop_last=False,
-        )
-        dataset_val = LightlyDataset.from_torch_dataset(
-            FakeData(
-                size=n_samples,
-                image_size=(3, 32, 32),
-                num_classes=num_classes,
-                transform=ToTensor(),
-                random_offset=10,
-            )
-        )
-        dataloader_val = DataLoader(
-            dataset_val,
-            batch_size=batch_size,
-            num_workers=0,
-            shuffle=False,
-            drop_last=False,
-        )
-
-        model = _BenchmarkDummyModel(
-            dataloader_kNN=dataloader_train, num_classes=num_classes
-        )
-
-        trainer = Trainer(
-            devices=n_devices,
-            accelerator="cpu",
-            strategy=DDPStrategy(find_unused_parameters=False),
-            max_epochs=10,
-        )
-        trainer.fit(
-            model,
-            train_dataloaders=dataloader_train,
-            val_dataloaders=dataloader_val,
-        )
-
-        assert model.max_accuracy == 0.953125
